Create type declaration file for sunburst
I am currently working with the Sunburst component for a university project, and because TypeScript is also an important part of that I wanted to create the type definitions.
I based this file on the type definitions of the other nivo components and I hope I did a well enough job here. Unlike most other nivo packages, sunburst is a bit lax when it comes to it's data property, so I worked a bit with optional fields in interfaces.