Add `useUTC` to types
After some messing around with Nivo and line charts, I had a issue where I discovered the dates I was passing in were being converted to the clients timezone, even though I wasn't supplying a timezone with the date and these dates I was passing in were actually already correct. To get around this we discovered a `useUTC` option which we set to false. 

Although it seems like its missing from the types.