fix(scales): fix timeScale min/max values and typings (#743)
* Support for min/max in the TimeScale.

* Min/max take same format as input data

If the input data is formatted and the format property of the scale is set - exp. '%Y-%m-%d' - the min/max properties provided explicitly were not taken into account. Instead the actual range was used - this was crashing as the values `values.max` and `values.min` are already converted to native format.