Support DerivedNodeProp<number> for passing nodeSize function to ResponsiveScatterPlot
Adds support for DerivedNodeProp<number> so when a function is passed to nodeSize
the id of the node can be referenced. The current implementation uses only DerivedDatumProp<number>
and the Datum type only has x and y fields where as Node also contains its id.

Also, the nivo docs at https://nivo.rocks/scatterplot/ specify
"it will receive the current node and must return a number"
which is consistent with the proposed type.