fix(bump): Return a new serie reference when color or styles change
This is the same fix as applied to area-bump in 412f0de9
Contrary to area-bump, the previous implementation didn't lead to a visible bug, as whenever e.g. someone hovers over the graph, the memoization of the rendering component is broken by another object anyway, hence it doesn't matter that from a React's perspective the series objects are still the same.

Therefore this commit can be seen as a preemptive measure to avoid future problems.