fix: Relax constraints on `@react-spring/web`
Having a constraint on `@react-spring/web` blocking it to a very precise version (here: `9.4.5`) can push users of nivo into pulling it several times in their node_modules or bundles if they don't really check for duplicated packages. Indeed it's rather simple to have two distinct libraries requesting it at version `9.x` at the same time, and if one of them (not nivo) request something not being `9.4.5` they we would pull the library twice.

It seems that given `@react-spring/web` seems to follow semver, there is not real risk to say nivo supports `^9.4.5` instead of `9.4.5` only.